+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 <!-- - Amélioration des performances lors du chargement des données -->
 - Alexnet model weight initialization for best convergence.
+- The dataset transformation for context encoding of JEPA is updated with image region masker.
 
 ### Fixed
 <!-- - Correction d'un bug empêchant l'export dans certains cas -->